backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
aiosqlite==0.20.0
python-multipart
